+++ src/main/java/com/tropo/jobportal/repository/JobPostRepository.java
@@ -3,5 +3,5 @@ package com.tropo.jobportal.repository;
 import com.tropo.jobportal.bean.JobPost;
 import org.springframework.data.repository.CrudRepository;
 
-public interface JobPostRepository extends CrudRepository<JobPost, Integer> {
+public interface JobPostRepository extends CrudRepository<JobPost, Long> {
 }
